changed some files to typescript and refactored code
