+++ frontend/src/components/dashboard/TopicsList/TopicsList.module.css
@@ -1,4 +1,4 @@
-.megaContainer{
+.megaContainer {
   padding-top: 10%;
   padding-bottom: 5%;
 }
@@ -21,37 +21,37 @@
 .card-bg-1 {
   background-color: #0f2027;
   background-image: linear-gradient(315deg, #0f2027 0%, #2c5364 74%);
-  color: #FFF;
+  color: #fff;
 }
 
 .card-bg-2 {
   background-color: #3a1c71;
   background-image: linear-gradient(315deg, #1c0448 0%, #401014 74%);
-  color: #FFF;
+  color: #fff;
 }
 
 .card-bg-3 {
   background-color: #232526;
   background-image: linear-gradient(315deg, #232526 0%, #414345 74%);
-  color: #FFF;
+  color: #fff;
 }
 
 .card-bg-4 {
   background-color: #403b4a;
   background-image: linear-gradient(315deg, #362d48 0%, #2d2f04 74%);
-  color: #FFF;
+  color: #fff;
 }
 
 .card-bg-5 {
   background-color: #36127d;
   background-image: linear-gradient(315deg, #113d08 0%, #2f0a33 74%);
-  color: #FFF;
+  color: #fff;
 }
 
 .card-bg-6 {
   background-color: #0c6579;
   background-image: linear-gradient(315deg, #453a35 0%, #54491b 74%);
-  color: #FFF;
+  color: #fff;
 }
 
 .modalClass .modal-dialog {
@@ -59,13 +59,11 @@
   max-width: 50rem;
 }
 
-
 .containerPadding {
-  padding-left: 150px; 
-  padding-right: 150px; 
+  padding-left: 150px;
+  padding-right: 150px;
   padding-top: 20px;
 }
-
 
 @media (max-width: 1300px) {
   .modalClass .modal-dialog {
@@ -73,36 +71,36 @@
   }
 
   .containerPadding {
-    padding-left: 20px; 
-    padding-right: 20px; 
+    padding-left: 20px;
+    padding-right: 20px;
     padding-top: 10px;
   }
 }
 
 @media (max-width: 900px) {
   .containerPadding {
-    padding-left: 10px; 
-    padding-right: 10px; 
+    padding-left: 10px;
+    padding-right: 10px;
     padding-top: 10px;
   }
 }
 
 @media (max-height: 1070px) {
-  .megaContainer{
+  .megaContainer {
     padding-top: 7.5%;
     padding-bottom: 5%;
   }
 }
 
 @media (max-height: 800px) {
-  .megaContainer{
+  .megaContainer {
     padding-top: 5%;
     padding-bottom: 5%;
   }
 }
 
 @media (max-height: 600px) {
-  .megaContainer{
+  .megaContainer {
     padding-top: -5%;
     padding-bottom: 5%;
   }
