+++ frontend/src/pages/ErrorPage.tsx
@@ -1,10 +1,9 @@
 import React from "react";
 
 const ErrorPage = () => {
-
   return (
     <div>
-        <h1>ErrorPage</h1>
+      <h1>ErrorPage</h1>
     </div>
   );
 };
