+++ frontend/src/pages/HomeRootPage.jsx
@@ -1,5 +1,11 @@
 import React from "react";
-import { Outlet, useLocation, useLoaderData, redirect, json } from "react-router-dom";
+import {
+  Outlet,
+  useLocation,
+  useLoaderData,
+  redirect,
+  json,
+} from "react-router-dom";
 import LoggedInHeader from "../components/headers/PrimaryHeader/LoggedInHeader/LoggedInHeader";
 
 const BaseURL = import.meta.env.VITE_API_BASE_URL;
@@ -9,23 +15,22 @@ const HomeRootPage = () => {
   const location = useLocation();
   const route = location.pathname;
 
-  if (route !== '/' && route !== '/login' && route !== '/signup'){
-    
+  if (route !== "/" && route !== "/login" && route !== "/signup") {
     return (
-      <>  
-          <LoggedInHeader username={data.username} />
-          <main>
-              <Outlet />
-          </main>
+      <>
+        <LoggedInHeader username={data.username} />
+        <main>
+          <Outlet />
+        </main>
       </>
-    );  
+    );
   }
 
   return (
     <>
-        <main>
-            <Outlet />
-        </main>
+      <main>
+        <Outlet />
+      </main>
     </>
   );
 };
@@ -33,26 +38,24 @@ const HomeRootPage = () => {
 export default HomeRootPage;
 
 export const loader = async ({ request, params }) => {
-
-  const token = localStorage.getItem('authToken');
+  const token = localStorage.getItem("authToken");
 
   const response = await fetch(`${BaseURL}/user/profile`, {
-    method: 'GET',
+    method: "GET",
     headers: {
-        'Authorization': 'Bearer ' + token
-    }
+      Authorization: "Bearer " + token,
+    },
   });
 
   if (!response.ok) {
-     
-      if (response.status === 401 || response.status === 403){
-          return redirect("/");
-      } 
+    if (response.status === 401 || response.status === 403) {
+      return redirect("/");
+    }
 
-      return json({ message: "User does not exist." }, { status: 500 });
-  } 
+    return json({ message: "User does not exist." }, { status: 500 });
+  }
 
   const user = await response.json();
 
   return user;
-}
+};
