+++ frontend/src/pages/SignupPage.jsx
@@ -6,26 +6,23 @@ import { deserializeUserDetails } from "../utils/parseJWT";
 const BaseURL = import.meta.env.VITE_API_BASE_URL;
 
 const SignupPage = () => {
-
-  return (
-    <SignUpForm />
-  );
+  return <SignUpForm />;
 };
 
 export default SignupPage;
 
 export const action = async ({ request, params }) => {
   const loginData = Object.fromEntries(await request.formData());
-  
+
   // check the password that the user enters for these two fields match before making request
-  if (loginData.password !== loginData["confirm-password"]){
+  if (loginData.password !== loginData["confirm-password"]) {
     return { errors: "Password and Confirm Password do not match.", code: 401 };
   }
 
-  // create request body object to submit and make the post request 
+  // create request body object to submit and make the post request
   const requestData = {
     username: loginData.username,
-    password: loginData.password
+    password: loginData.password,
   };
 
   const response = await fetch(`${BaseURL}/user/register`, {
@@ -36,7 +33,7 @@ export const action = async ({ request, params }) => {
     body: JSON.stringify(requestData),
   });
 
-  if (!response.ok){
+  if (!response.ok) {
     if (response.status === 400) {
       const errdata = await response.json();
 
@@ -58,7 +55,7 @@ export const action = async ({ request, params }) => {
   const token = resData.accessToken;
 
   localStorage.setItem("authToken", token);
-  
+
   // when the user creates a new account it is their first time logging in so redirect to the questionnaire page
   return redirect("/home/questionnaire");
 };
